Moved some of the install processing to the NAS Proxy RPM file.
The RPM file should contains all of the processing necessary to install and start the NAS Proxy.  That work should NOT be done from the kickstart processing.

This sets the stags for having the same requirement for the NAS Encryptor.  The NAS Encryptor should have an RPM file that does everything necessary to install and start the NAS Encryptor.  It shouldn't have to start from the kickstart operations.