Deploy to AWS updates.
It still doesn't do everything, but there are notes in the file that explain a lot of the stuff that needs to be automated.  So the steps are mostly there.

Currently the end result would be that the OVA file is imported and turned into an AWS AMI.  There aren't steps to "Launch" it yet.  We'll worry about that after we get it to the point where we can automate the process of uploading and importing the NASProxy.ova to an AWS AMI.